pyproject.toml
[tool.poetry]
name = "py2mac"
version = "0.0.0.post20.dev0+b7b2042"
description = ""
authors = ["Rafal Wytrykus <rafal@wytrykus.dev>"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.11"
pyobjc = "^10.2"
pyobjc-framework-ApplicationServices = "^10.2"
pyobjc-framework-Accessibility = "^10.2"
attrs = "^23.2.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
black = "^24.4.0"
ruff = "^0.4.1"
pytest = "^8.1.1"
isort = "^5.13.2"
fire = "^0.6.0"
mypy = "^1.9.0"
ipython = "^8.23.0"
ipywidgets = "^8.1.5"
chainlit = "^1.3.1"
sphinx = "^8.1.3"
sphinx-autodoc-typehints = "^2.5.0"
sphinx-rtd-theme = "^3.0.1"
langchain = "^0.3.4"
langchain-openai = "^0.2.4"


[tool.poetry-dynamic-versioning]
enable = true
pattern = "default-unprefixed"
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"


[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.ruff]
exclude = ["build", "dist", ".venv", ".tox", ".mypy_cache", ".pytest_cache", ".ruff_cache", ".ruff"]
lint.extend-select = ["I"]

[tool.pytest]
addopts = ["--color=yes", "--verbose"]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
ignore_missing_imports = true
